+++ src/certification-tools/tools.ts
@@ -1,42 +1,21 @@
-import { UbirchCertification } from '../certification/certification';
-import i18n from '../utils/translations';
-import { decode, encode } from '@msgpack/msgpack';
-import * as base45 from 'base45';
+import UbirchProtocol from '@ubirch/ubirch-protocol-verifier/src/upp';
 import { Buffer } from 'buffer';
-import { createHash } from 'crypto';
-import zlib from 'zlib';
-import { EError, EUppStates, EUppTypes, IUbirchError, IUbirchSignedCertificationResponse, IUbirchUpp } from '../models/models';
+import { EError, IUbirchError, IUbirchSignedCertificationResponse } from '../models/models';
+import i18n from '../utils/translations';
 
 export class UbirchCertificationTools {
   protected static ProtocolVersion = 2;
 
-// Commented out as it is currently not used
-// const PLAIN = (ProtocolVersion << 4) | 0x01
-  private static SIGNED = (UbirchCertificationTools.ProtocolVersion << 4) | 0x02;
-  private static CHAINED = (UbirchCertificationTools.ProtocolVersion << 4) | 0x03;
-  public static UPP_TYPE = {
-    CHAINED: 0x00,
-    SIGNED: 0xEE
-  }
-
-  private static CERT_PREFIX = "C01:";
-
   public static getMsgPackPayload (jsonPayload: any): Uint8Array {
-    return encode(jsonPayload);
+    return UbirchProtocol.tools.createMsgPackPayloadFromJSON(jsonPayload);
   }
 
   public static getHashedPayload (payload): string {
-    return createHash('sha256').update(payload).digest('base64');
+    return UbirchProtocol.tools.getHashedPayload(payload);
   }
 
   public static replaceHashByMsgPackInUpp(hashUpp: Buffer, msgPackPayload: Uint8Array) {
-    let unpacked_upp: any[] = decode(hashUpp) as any[];
-    const uppLength = unpacked_upp.length;
-
-    unpacked_upp[uppLength - 2] = msgPackPayload;
-    unpacked_upp[uppLength - 3] = UbirchCertificationTools.UPP_TYPE.SIGNED;
-
-    return encode(unpacked_upp);
+    return UbirchProtocol.tools.replaceHashByMsgPackInUpp(hashUpp, msgPackPayload);
   }
 
   /**
@@ -45,15 +24,11 @@ export class UbirchCertificationTools {
    * @returns {string} signed UPP
    */
   public static packSignedUpp(msgPackUpp: Uint8Array): string {
-    const buf = Buffer.from(msgPackUpp);
-    const zlibbed_upp = zlib.deflateSync(buf);
-
-    const base45ed_upp = base45.encode(zlibbed_upp);
-
-    return UbirchCertificationTools.CERT_PREFIX + base45ed_upp;
+    return UbirchProtocol.tools.packSignedUpp(msgPackUpp);
   }
 
   public static extractSignedUpp(resultObj: IUbirchSignedCertificationResponse): Buffer {
+
     const data = resultObj?.data?.body;
     if (!data) {
       throw { code: EError.UNKNOWN_ERROR } as IUbirchError;
@@ -71,7 +46,7 @@ export class UbirchCertificationTools {
       throw { code: EError.CERTIFICATION_FAILED_NO_UPP } as IUbirchError;
     }
 
-    return Buffer.from(resultObj.data.body.upp, 'base64');
+    return UbirchProtocol.tools.unpackBase64String(resultObj.data.body.upp);
   }
 }